+++ src/mut_bump_vec.rs
@@ -900,6 +900,8 @@ impl<T, A: MutBumpAllocator> MutBumpVec<T, A> {
     ///
     /// This is behaviorally identical to [`FromIterator::from_iter`].
     ///
+    /// If you have an `impl ExactSizeIterator` then you can use [`from_iter_exact_in`](Self::from_iter_exact_in) instead for better performance.
+    ///
     /// # Panics
     /// Panics if the allocation fails.
     ///
@@ -923,6 +925,8 @@ impl<T, A: MutBumpAllocator> MutBumpVec<T, A> {
     /// Create a new [`MutBumpVec`] whose elements are taken from an iterator and allocated in the given `bump`.
     ///
     /// This is behaviorally identical to [`FromIterator::from_iter`].
+    ///
+    /// If you have an `impl ExactSizeIterator` then you can use [`try_from_iter_exact_in`](Self::try_from_iter_exact_in) instead for better performance.
     ///
     /// # Errors
     /// Errors if the allocation fails.
@@ -955,6 +959,77 @@ impl<T, A: MutBumpAllocator> MutBumpVec<T, A> {
 
         for value in iter {
             vec.generic_push(value)?;
+        }
+
+        Ok(vec)
+    }
+
+    /// Create a new [`MutBumpVec`] whose elements are taken from an iterator and allocated in the given `bump`.
+    ///
+    /// This is just like [`from_iter_in`](Self::from_iter_in) but optimized for an [`ExactSizeIterator`].
+    ///
+    /// # Panics
+    /// Panics if the allocation fails.
+    ///
+    /// # Examples
+    /// ```
+    /// # use bump_scope::{Bump, MutBumpVec};
+    /// # let mut bump: Bump = Bump::new();
+    /// let vec = MutBumpVec::from_iter_exact_in([1, 2, 3], &mut bump);
+    /// assert_eq!(vec, [1, 2, 3]);
+    /// ```
+    #[must_use]
+    #[inline(always)]
+    #[cfg(feature = "panic-on-alloc")]
+    pub fn from_iter_exact_in<I>(iter: I, allocator: A) -> Self
+    where
+        I: IntoIterator<Item = T>,
+        I::IntoIter: ExactSizeIterator,
+    {
+        panic_on_error(Self::generic_from_iter_exact_in(iter, allocator))
+    }
+
+    /// Create a new [`MutBumpVec`] whose elements are taken from an iterator and allocated in the given `bump`.
+    ///
+    /// This is just like [`from_iter_in`](Self::from_iter_in) but optimized for an [`ExactSizeIterator`].
+    ///
+    /// # Errors
+    /// Errors if the allocation fails.
+    ///
+    /// # Examples
+    /// ```
+    /// # use bump_scope::{Bump, MutBumpVec};
+    /// # let mut bump: Bump = Bump::try_new()?;
+    /// let vec = MutBumpVec::try_from_iter_exact_in([1, 2, 3], &mut bump)?;
+    /// assert_eq!(vec, [1, 2, 3]);
+    /// # Ok::<(), bump_scope::alloc::AllocError>(())
+    /// ```
+    #[inline(always)]
+    pub fn try_from_iter_exact_in<I>(iter: I, allocator: A) -> Result<Self, AllocError>
+    where
+        I: IntoIterator<Item = T>,
+        I::IntoIter: ExactSizeIterator,
+    {
+        Self::generic_from_iter_exact_in(iter, allocator)
+    }
+
+    #[inline]
+    pub(crate) fn generic_from_iter_exact_in<E: ErrorBehavior, I>(iter: I, allocator: A) -> Result<Self, E>
+    where
+        I: IntoIterator<Item = T>,
+        I::IntoIter: ExactSizeIterator,
+    {
+        let mut iter = iter.into_iter();
+        let len = iter.len();
+
+        let mut vec = Self::generic_with_capacity_in(len, allocator)?;
+
+        while vec.len() != vec.capacity() {
+            match iter.next() {
+                // SAFETY: we checked above that `len != capacity`, so there is space
+                Some(value) => unsafe { vec.push_unchecked(value) },
+                None => break,
+            }
         }
 
         Ok(vec)